// src/interfaces/system.ts
import {Request} from "express";

/**
 Author: Revelation.AF
 Engine: [//]
 Git: nusktec
 **/
//interfaces
interface EmailCategoryInterface {
    PASSWORD_RESET: string;
    ORDER_CONFIRMATION: string;
    SHIPPING_NOTIFICATION: string;
    ACCOUNT_CREATION: string;
    INVOICE: string;
    PROMOTIONAL: string;
    NEWSLETTER: string;
    EVENT_INVITATION: string;
    PRODUCT_LAUNCH: string;
    ABANDONED_CART: string;
    CUSTOMER_SUPPORT: string;
    FEEDBACK_SURVEYS: string;
    ISSUE_RESOLUTION: string;
    ACCOUNT_ALERTS: string;
    SYSTEM_ALERTS: string;
    USAGE_REPORTS: string;
    SUBSCRIPTION_RENEWAL: string;
    PRIVACY_POLICY_UPDATE: string;
    GDPR_COMPLIANCE: string;
    TEAM_ANNOUNCEMENTS: string;
    MEETING_INVITATIONS: string;
    NEW_FOLLOWER_CONNECTION: string;
    COMMENT_MENTION: string;
    FRIEND_REQUEST: string;
    GENERAL: string
}

type EmailCategoryType =
    | "Password Reset"
    | "Order Confirmation"
    | "Shipping Notification"
    | "Account Creation"
    | "Invoice"
    | "Promotional"
    | "Newsletter"
    | "Event Invitation"
    | "Product Launch"
    | "Abandoned Cart"
    | "Customer Support"
    | "Feedback/Surveys"
    | "Issue Resolution"
    | "Account Alerts"
    | "System Alerts"
    | "Usage Reports"
    | "Subscription Renewal"
    | "Privacy Policy Update"
    | "GDPR Compliance"
    | "Team Announcements"
    | "Meeting Invitations"
    | "New Follower/Connection"
    | "Comment or Mention"
    | "General Information"
    | "Friend Request";

const HttpStatusCode = Object.freeze({
    // 1xx Informational Responses
    CONTINUE: 100,
    SWITCHING_PROTOCOLS: 101,
    PROCESSING: 102, // WebDAV
    EARLY_HINTS: 103,

    // 2xx Success
    OK: 200,
    CREATED: 201,
    ACCEPTED: 202,
    NON_AUTHORITATIVE_INFORMATION: 203,
    NO_CONTENT: 204,
    RESET_CONTENT: 205,
    PARTIAL_CONTENT: 206,
    MULTI_STATUS: 207, // WebDAV
    ALREADY_REPORTED: 208, // WebDAV
    IM_USED: 226,

    // 3xx Redirection
    MULTIPLE_CHOICES: 300,
    MOVED_PERMANENTLY: 301,
    FOUND: 302,
    SEE_OTHER: 303,
    NOT_MODIFIED: 304,
    USE_PROXY: 305,
    SWITCH_PROXY: 306,
    TEMPORARY_REDIRECT: 307,
    PERMANENT_REDIRECT: 308,

    // 4xx Client Errors
    BAD_REQUEST: 400,
    UNAUTHORIZED: 401,
    PAYMENT_REQUIRED: 402,
    FORBIDDEN: 403,
    NOT_FOUND: 404,
    METHOD_NOT_ALLOWED: 405,
    NOT_ACCEPTABLE: 406,
    PROXY_AUTHENTICATION_REQUIRED: 407,
    REQUEST_TIMEOUT: 408,
    CONFLICT: 409,
    GONE: 410,
    LENGTH_REQUIRED: 411,
    PRECONDITION_FAILED: 412,
    PAYLOAD_TOO_LARGE: 413,
    URI_TOO_LONG: 414,
    UNSUPPORTED_MEDIA_TYPE: 415,
    RANGE_NOT_SATISFIABLE: 416,
    EXPECTATION_FAILED: 417,
    IM_A_TEAPOT: 418,
    MISDIRECTED_REQUEST: 421,
    UNPROCESSABLE_ENTITY: 422, // WebDAV
    LOCKED: 423, // WebDAV
    FAILED_DEPENDENCY: 424, // WebDAV
    TOO_EARLY: 425,
    UPGRADE_REQUIRED: 426,
    PRECONDITION_REQUIRED: 428,
    TOO_MANY_REQUESTS: 429,
    REQUEST_HEADER_FIELDS_TOO_LARGE: 431,
    UNAVAILABLE_FOR_LEGAL_REASONS: 451,

    // 5xx Server Errors
    INTERNAL_SERVER_ERROR: 500,
    NOT_IMPLEMENTED: 501,
    BAD_GATEWAY: 502,
    SERVICE_UNAVAILABLE: 503,
    GATEWAY_TIMEOUT: 504,
    HTTP_VERSION_NOT_SUPPORTED: 505,
    VARIANT_ALSO_NEGOTIATES: 506,
    INSUFFICIENT_STORAGE: 507, // WebDAV
    LOOP_DETECTED: 508, // WebDAV
    NOT_EXTENDED: 510,
    NETWORK_AUTHENTICATION_REQUIRED: 511,
} as const);

type HttpStatusCode = typeof HttpStatusCode[keyof typeof HttpStatusCode];

export interface AuthRequest extends Request {
    auth?: any; // Add user data to the request object
}

export {EmailCategoryType, EmailCategoryInterface, HttpStatusCode}